use toHexString instead of decodeString to ensure IIDs are not merged in export
cleanup error message

WIP using rocksdb as id mapping store

use new valid exported test file

cleanup

fix checkstyle

refactor Importer

cleanup

address first comments

delete interface

cleanup

WIP include comments as TODOs

fix ignored data

rename
