+++ .gitignore
@@ -36,7 +36,6 @@ bazel-*
 .classpath
 
 # Compiled
-data/
 dist/
 out/
 *.class
