+++ common/test/logback-test.xml
@@ -32,7 +32,7 @@
     WARN:  (Unexpected events that are not recommended to have happened)
     ERROR: (Incorrect events that disrupts the server operation, against user expectation)
     -->
-    <logger name="com.vaticle.typedb.core" level="INFO">
+    <logger name="com.vaticle.typedb.core" level="DEBUG">
         <appender-ref ref="STDOUT"/>
     </logger>
 </configuration>
