+++ migrator/MigratorClient.java
@@ -20,8 +20,6 @@ package com.vaticle.typedb.core.migrator;
 
 import com.vaticle.typedb.core.common.exception.ErrorMessage;
 import com.vaticle.typedb.core.common.exception.TypeDBException;
-import com.vaticle.typedb.core.migrator.proto.MigratorGrpc;
-import com.vaticle.typedb.core.migrator.proto.MigratorProto;
 import io.grpc.ManagedChannel;
 import io.grpc.ManagedChannelBuilder;
 import io.grpc.stub.StreamObserver;
@@ -35,51 +33,86 @@ import java.util.concurrent.CountDownLatch;
 //       and it should be able to stream import/export file to/from the server
 public class MigratorClient {
 
-    private final MigratorGrpc.MigratorStub streamingStub;
+    private final MigratorGrpc.MigratorStub stub;
 
     public MigratorClient(int serverPort) {
         String uri = "localhost:" + serverPort;
         ManagedChannel channel = ManagedChannelBuilder.forTarget(uri).usePlaintext().build();
-        streamingStub = MigratorGrpc.newStub(channel);
+        stub = MigratorGrpc.newStub(channel);
     }
 
     public boolean importData(String database, String filename, Map<String, String> remapLabels) {
-        MigratorProto.ImportData.Req req = MigratorProto.ImportData.Req.newBuilder()
+        MigratorProto.Import.Req req = MigratorProto.Import.Req.newBuilder()
                 .setDatabase(database)
                 .setFilename(filename)
                 .putAllRemapLabels(remapLabels)
                 .build();
-        ResponseObserver streamObserver = new ResponseObserver(new ProgressPrinter("import"));
-        streamingStub.importData(req, streamObserver);
+        ResponseObserver.Import streamObserver = new ResponseObserver.Import(new ProgressPrinter.Import());
+        stub.importData(req, streamObserver);
         streamObserver.await();
         return streamObserver.success();
     }
 
     public boolean exportData(String database, String filename) {
-        MigratorProto.ExportData.Req req = MigratorProto.ExportData.Req.newBuilder()
+        MigratorProto.Export.Req req = MigratorProto.Export.Req.newBuilder()
                 .setDatabase(database)
                 .setFilename(filename)
                 .build();
-        ResponseObserver streamObserver = new ResponseObserver(new ProgressPrinter("export"));
-        streamingStub.exportData(req, streamObserver);
+        ResponseObserver.Export streamObserver = new ResponseObserver.Export(new ProgressPrinter.Export());
+        stub.exportData(req, streamObserver);
         streamObserver.await();
         return streamObserver.success();
     }
 
-    static class ResponseObserver implements StreamObserver<MigratorProto.Job.Res> {
+    static abstract class ResponseObserver<T> implements StreamObserver<T> {
 
-        private final ProgressPrinter progressPrinter;
         private final CountDownLatch latch;
         private boolean success;
 
-        public ResponseObserver(ProgressPrinter progressPrinter) {
-            this.progressPrinter = progressPrinter;
+        public ResponseObserver() {
             this.latch = new CountDownLatch(1);
         }
 
-        @Override
-        public void onNext(MigratorProto.Job.Res res) {
-            progressPrinter.onProgress(ProgressPrinter.Progress.of(res.getProgress()));
+        private static class Import extends ResponseObserver<MigratorProto.Import.Progress> {
+
+            private final ProgressPrinter.Import progressPrinter;
+
+            public Import(ProgressPrinter.Import progressPrinter) {
+                super();
+                this.progressPrinter = progressPrinter;
+            }
+
+            @Override
+            public void onNext(MigratorProto.Import.Progress progress) {
+                progressPrinter.onProgress(progress);
+            }
+
+            @Override
+            public void onCompleted() {
+                super.onCompleted();
+                progressPrinter.onCompleted();
+            }
+        }
+
+        private static class Export extends ResponseObserver<MigratorProto.Export.Progress> {
+
+            private final ProgressPrinter.Export progressPrinter;
+
+            public Export(ProgressPrinter.Export progressPrinter) {
+                super();
+                this.progressPrinter = progressPrinter;
+            }
+
+            @Override
+            public void onNext(MigratorProto.Export.Progress progress) {
+                progressPrinter.onProgress(progress);
+            }
+
+            @Override
+            public void onCompleted() {
+                super.onCompleted();
+                progressPrinter.onCompleted();
+            }
         }
 
         @Override
@@ -91,7 +124,6 @@ public class MigratorClient {
 
         @Override
         public void onCompleted() {
-            progressPrinter.onCompletion();
             success = true;
             latch.countDown();
         }
@@ -109,25 +141,21 @@ public class MigratorClient {
         }
     }
 
-    private static class ProgressPrinter {
+    private static abstract class ProgressPrinter {
 
         private static final String[] ANIM = new String[]{"-", "\\", "|", "/"};
         private static final String STATUS_STARTING = "starting";
         private static final String STATUS_IN_PROGRESS = "in progress";
         private static final String STATUS_COMPLETED = "completed";
 
-        private final String type;
         private final Timer timer = new Timer();
 
-        private String status = STATUS_STARTING;
-
-        private Progress progress;
+        String status = STATUS_STARTING;
 
         private int anim = 0;
         private int lines = 0;
 
-        public ProgressPrinter(String type) {
-            this.type = type;
+        public ProgressPrinter() {
             TimerTask task = new TimerTask() {
                 @Override
                 public void run() {
@@ -137,24 +165,23 @@ public class MigratorClient {
             timer.scheduleAtFixedRate(task, 0, 100);
         }
 
-        public void onProgress(Progress progress) {
-            status = STATUS_IN_PROGRESS;
-            this.progress = progress;
-        }
-
-        public void onCompletion() {
+        public void onCompleted() {
             status = STATUS_COMPLETED;
             step();
             timer.cancel();
         }
 
+        abstract String type();
+
+        abstract String formattedProgress();
+
         private synchronized void step() {
             StringBuilder builder = new StringBuilder();
-            builder.append(String.format("$x isa %s,\n    has status \"%s\";", type, status));
+            builder.append(String.format("$x isa %s,\n    has status \"%s\";", type(), status));
 
             if (!status.equals(STATUS_STARTING)) {
                 builder.append("\n\n");
-                builder.append(progress.toString());
+                builder.append(formattedProgress());
                 builder.append(";");
                 anim = (anim + 1) % ANIM.length;
                 builder.append(" ").append(ANIM[anim]);
@@ -166,86 +193,83 @@ public class MigratorClient {
             lines = output.split("\n").length;
         }
 
+        private static class Import extends ProgressPrinter {
 
-        static abstract class Progress {
+            private MigratorProto.Import.Progress prog;
 
-            public static Progress of(MigratorProto.Job.Progress progress) {
-                if (progress.hasExportProgress()) return Export.of(progress.getExportProgress());
-                else return Import.of(progress.getImportProgress());
+            void onProgress(MigratorProto.Import.Progress progress) {
+                this.status = STATUS_IN_PROGRESS;
+                this.prog = progress;
             }
 
-            static class Import extends Progress {
-
-                private final MigratorProto.Job.ImportProgress prog;
-
-                private Import(MigratorProto.Job.ImportProgress progress) {
-                    this.prog = progress;
-                }
-
-                static Import of(MigratorProto.Job.ImportProgress importProgress) {
-                    return new Import(importProgress);
-                }
-
-                @Override
-                public String toString() {
-                    StringBuilder progressStr = new StringBuilder();
-                    progressStr.append(prog.getAttributes() == 0 ? String.format("Attribute: %d", prog.getAttributesCurrent()) :
-                            String.format("Attribute: %d/%d (%.1f%%)", prog.getAttributesCurrent(), prog.getAttributes(), 100.0 * prog.getAttributesCurrent() / prog.getAttributes()));
-                    long currentThings = prog.getAttributesCurrent();
-                    long things = prog.getAttributes();
-                    if (!prog.getInitialising()) {
-                        progressStr.append(" ");
-                        progressStr.append(prog.getOwnerships() == 0 ? String.format("[ ownership: %d/%d ]", prog.getOwnershipsCurrent(), prog.getOwnerships()) :
-                                String.format("[ ownership: %d/%d (%.1f%%) ]", prog.getOwnershipsCurrent(), prog.getOwnerships(), 100.0 * prog.getOwnershipsCurrent() / prog.getOwnerships()));
-                        progressStr.append("\n");
-                        progressStr.append(prog.getEntities() == 0 ? String.format("Entity: %d/%d", prog.getEntitiesCurrent(), prog.getEntities()) :
-                                String.format("Entity: %d/%d (%.1f%%)", prog.getEntitiesCurrent(), prog.getEntities(), 100.0 * prog.getEntitiesCurrent() / prog.getEntities()));
-                        progressStr.append("\n");
-                        progressStr.append(prog.getRelations() == 0 ? String.format("Relation: %d/%d", prog.getRelationsCurrent(), prog.getRelations()) :
-                                String.format("Relation: %d/%d (%.1f%%)", prog.getRelationsCurrent(), prog.getRelations(), 100.0 * prog.getRelationsCurrent() / prog.getRelations()));
-                        progressStr.append(" ");
-                        progressStr.append(prog.getRoles() == 0 ? String.format("[ role: %d/%d ]", prog.getRolesCurrent(), prog.getRoles()) :
-                                String.format("[ role: %d/%d (%.1f%%) ]", prog.getRolesCurrent(), prog.getRoles(), 100.0 * prog.getRolesCurrent() / prog.getRoles()));
-                        currentThings += prog.getEntitiesCurrent() + prog.getRelationsCurrent() + prog.getOwnershipsCurrent() + prog.getRolesCurrent();
-                        things += prog.getEntities() + prog.getRelations() + prog.getOwnerships() + prog.getRoles();
-
-                        progressStr.append("\n");
-                        progressStr.append(String.format("Total: %d/%d (%.1f%%)", currentThings, things, 100.0 * currentThings / things));
-                    }
-                    return progressStr.toString();
-                }
+            @Override
+            String formattedProgress() {
+                StringBuilder progressStr = new StringBuilder();
+                progressStr.append(prog.getAttributes() == 0 ?
+                        String.format("Attribute: %d", prog.getAttributesCurrent()) :
+                        String.format("Attribute: %d/%d (%.1f%%)", prog.getAttributesCurrent(), prog.getAttributes(),
+                                100.0 * prog.getAttributesCurrent() / prog.getAttributes()));
+                progressStr.append("\n");
+                progressStr.append(prog.getEntities() == 0 ?
+                        String.format("Entity: %d", prog.getEntitiesCurrent()) :
+                        String.format("Entity: %d/%d (%.1f%%)", prog.getEntitiesCurrent(), prog.getEntities(),
+                                100.0 * prog.getEntitiesCurrent() / prog.getEntities()));
+                progressStr.append("\n");
+                progressStr.append(prog.getRelations() == 0 ?
+                        String.format("Relation: %d", prog.getRelationsCurrent()) :
+                        String.format("Relation: %d/%d (%.1f%%)", prog.getRelationsCurrent(), prog.getRelations(),
+                                100.0 * prog.getRelationsCurrent() / prog.getRelations()));
+                progressStr.append("\n");
+                long currentThings = prog.getAttributesCurrent() + prog.getEntitiesCurrent() + prog.getRelationsCurrent();
+                long things = prog.getAttributes() + prog.getEntities() + prog.getRelations();
+                progressStr.append("\n");
+                progressStr.append(String.format("Total: %d/%d (%.1f%%)", currentThings, things, 100.0 * currentThings / things));
+                return progressStr.toString();
             }
 
-            static class Export extends Progress {
+            @Override
+            String type() {
+                return "Import";
+            }
+        }
 
-                private final MigratorProto.Job.ExportProgress prog;
+        private static class Export extends ProgressPrinter {
 
-                public Export(MigratorProto.Job.ExportProgress exportProgress) {
-                    this.prog = exportProgress;
-                }
+            private MigratorProto.Export.Progress prog;
 
-                static Export of(MigratorProto.Job.ExportProgress exportProgress) {
-                    return new Export(exportProgress);
-                }
+            void onProgress(MigratorProto.Export.Progress progress) {
+                this.status = STATUS_IN_PROGRESS;
+                this.prog = progress;
+            }
 
-                @Override
-                public String toString() {
-                    StringBuilder progressStr = new StringBuilder();
-                    progressStr.append(prog.getAttributes() == 0 ? String.format("Attribute: %dd", prog.getAttributesCurrent(), prog.getAttributes()) :
-                            String.format("Attribute: %d/%d (%.1f%%)", prog.getAttributesCurrent(), prog.getAttributes(), 100.0 * prog.getAttributesCurrent() / prog.getAttributes()));
-                    progressStr.append("\n");
-                    progressStr.append(prog.getEntities() == 0 ? String.format("Entity: %dd", prog.getEntitiesCurrent(), prog.getEntities()) :
-                            String.format("Entity: %d/%d (%.1f%%)", prog.getEntitiesCurrent(), prog.getEntities(), 100.0 * prog.getEntitiesCurrent() / prog.getEntities()));
-                    progressStr.append("\n");
-                    progressStr.append(prog.getRelations() == 0 ? String.format("Relation: %d", prog.getRelationsCurrent(), prog.getRelations()) :
-                            String.format("Relation: %d/%d (%.1f%%)", prog.getRelationsCurrent(), prog.getRelations(), 100.0 * prog.getRelationsCurrent() / prog.getRelations()));
-                    progressStr.append("\n");
-                    long currentThings = prog.getAttributesCurrent() + prog.getEntitiesCurrent() + prog.getRelationsCurrent();
-                    long things = prog.getAttributes() + prog.getEntities() + prog.getRelations();
-                    progressStr.append("\n");
-                    progressStr.append(String.format("\nTotal: %d/%d (%.1f%%)", currentThings, things, 100.0 * currentThings / things));
-                    return progressStr.toString();
-                }
+            @Override
+            String formattedProgress() {
+                StringBuilder progressStr = new StringBuilder();
+                progressStr.append(prog.getAttributes() == 0 ?
+                        String.format("Attribute: %d", prog.getAttributesCurrent()) :
+                        String.format("Attribute: %d/%d (%.1f%%)", prog.getAttributesCurrent(), prog.getAttributes(),
+                                100.0 * prog.getAttributesCurrent() / prog.getAttributes()));
+                progressStr.append("\n");
+                progressStr.append(prog.getEntities() == 0 ?
+                        String.format("Entity: %d", prog.getEntitiesCurrent()) :
+                        String.format("Entity: %d/%d (%.1f%%)", prog.getEntitiesCurrent(), prog.getEntities(),
+                                100.0 * prog.getEntitiesCurrent() / prog.getEntities()));
+                progressStr.append("\n");
+                progressStr.append(prog.getRelations() == 0 ?
+                        String.format("Relation: %d", prog.getRelationsCurrent()) :
+                        String.format("Relation: %d/%d (%.1f%%)", prog.getRelationsCurrent(), prog.getRelations(),
+                                100.0 * prog.getRelationsCurrent() / prog.getRelations()));
+                progressStr.append("\n");
+                long currentThings = prog.getAttributesCurrent() + prog.getEntitiesCurrent() + prog.getRelationsCurrent();
+                long things = prog.getAttributes() + prog.getEntities() + prog.getRelations();
+                progressStr.append("\n");
+                progressStr.append(String.format("Total: %d/%d (%.1f%%)", currentThings, things, 100.0 * currentThings / things));
+                return progressStr.toString();
+            }
+
+            @Override
+            String type() {
+                return "Export";
             }
         }
     }
