+++ migrator/MigratorService.java
@@ -17,10 +17,8 @@
 
 package com.vaticle.typedb.core.migrator;
 
-import com.vaticle.typedb.core.TypeDB;
-import com.vaticle.typedb.core.common.exception.TypeDBException;
-import com.vaticle.typedb.core.migrator.proto.MigratorGrpc;
-import com.vaticle.typedb.core.migrator.proto.MigratorProto;
+import com.vaticle.typedb.core.migrator.data.DataExporter;
+import com.vaticle.typedb.core.migrator.data.DataImporter;
 import com.vaticle.typedb.core.rocks.RocksTypeDB;
 import io.grpc.Status;
 import io.grpc.StatusRuntimeException;
@@ -28,12 +26,8 @@ import io.grpc.stub.StreamObserver;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import java.nio.file.Path;
 import java.nio.file.Paths;
 import java.util.concurrent.CompletableFuture;
-import java.util.concurrent.ExecutionException;
-
-import static com.vaticle.typedb.core.common.exception.ErrorMessage.Internal.UNEXPECTED_INTERRUPTION;
 
 public class MigratorService extends MigratorGrpc.MigratorImplBase {
 
@@ -47,34 +41,39 @@ public class MigratorService extends MigratorGrpc.MigratorImplBase {
     }
 
     @Override
-    public void exportData(MigratorProto.ExportData.Req request, StreamObserver<MigratorProto.Job.Res> responseObserver) {
+    public void exportData(MigratorProto.Export.Req request, StreamObserver<MigratorProto.Export.Progress> responseObserver) {
         DataExporter exporter = new DataExporter(typedb, request.getDatabase(), Paths.get(request.getFilename()), version);
-        runMigrator(exporter, responseObserver);
-    }
-
-    @Override
-    public void importData(MigratorProto.ImportData.Req request, StreamObserver<MigratorProto.Job.Res> responseObserver) {
-        Path file = Paths.get(request.getFilename());
-        DataImporter importer = new DataImporter(typedb, request.getDatabase(), file, request.getRemapLabelsMap(), version);
-        runMigrator(importer, responseObserver);
-    }
-
-    private void runMigrator(Migrator migrator, StreamObserver<MigratorProto.Job.Res> responseObserver) {
         try {
-            CompletableFuture<Void> migratorJob = CompletableFuture.runAsync(migrator::run);
+            CompletableFuture<Void> migratorJob = CompletableFuture.runAsync(exporter::run);
             while (!migratorJob.isDone()) {
                 Thread.sleep(1000);
-                responseObserver.onNext(MigratorProto.Job.Res.newBuilder().setProgress(migrator.getProgress()).build());
+                responseObserver.onNext(exporter.getProgress());
             }
             migratorJob.get();
             responseObserver.onCompleted();
-        } catch (InterruptedException | ExecutionException e) {
-            throw TypeDBException.of(UNEXPECTED_INTERRUPTION);
+        } catch (Throwable e) {
+            LOG.error(e.getMessage(), e);
+            responseObserver.onError(exception(e));
+        }
+    }
+
+    @Override
+    public void importData(MigratorProto.Import.Req request, StreamObserver<MigratorProto.Import.Progress> responseObserver) {
+        DataImporter importer = new DataImporter(typedb, request.getDatabase(), Paths.get(request.getFilename()),
+                request.getRemapLabelsMap(), version);
+        try {
+            CompletableFuture<Void> migratorJob = CompletableFuture.runAsync(importer::run);
+            while (!migratorJob.isDone()) {
+                Thread.sleep(1000);
+                responseObserver.onNext(importer.getProgress());
+            }
+            migratorJob.get();
+            responseObserver.onCompleted();
         } catch (Throwable e) {
             LOG.error(e.getMessage(), e);
             responseObserver.onError(exception(e));
         } finally {
-            migrator.close();
+            importer.close();
         }
     }
 
@@ -82,7 +81,7 @@ public class MigratorService extends MigratorGrpc.MigratorImplBase {
         if (e instanceof StatusRuntimeException) {
             return (StatusRuntimeException) e;
         } else {
-            return Status.INTERNAL.withDescription(e.getMessage() + " Please check server logs for the stack trace.").asRuntimeException();
+            return Status.INTERNAL.withDescription(e.getMessage()).asRuntimeException();
         }
     }
 }
